requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
libvirt-python==9.8.0
pydantic==2.5.0